Required to merge f2py-only build process into branch

+++ source/fortran/fwbs_variables.f90
@@ -14,43 +14,43 @@ module fwbs_variables
 
   public
 
-  real(dp) :: bktlife
+  real(8) :: bktlife
   !! Full power blanket lifetime (years)
 
-  real(dp) :: coolmass
+  real(8) :: coolmass
   !! mass of water coolant (in shield, blanket, first wall, divertor) (kg)
 
-  real(dp) :: vvmass
+  real(8) :: vvmass
   !! vacuum vessel mass (kg)
 
-  real(dp) :: denstl
+  real(8) :: denstl
   !! density of steel (kg/m3)
   !#TODO: should this be in constants. Is currently an input. Should be a list of preapproved options?
 
-  real(dp) :: denw
+  real(8) :: denw
   !! density of tungsten (kg/m3)
   !#TODO: same as above with steel?
 
-  real(dp) :: dewmkg
+  real(8) :: dewmkg
   !! total mass of vacuum vessel + cryostat (kg) (calculated if blktmodel>0)
   !# TODO: blktmodel needs consolidating with iblanket
 
-  real(dp) :: emult
+  real(8) :: emult
   !! energy multiplication in blanket and shield
 
-  real(dp) :: emultmw
+  real(8) :: emultmw
   !! power due to energy multiplication in blanket and shield [MW]
 
-  real(dp) :: fblss
+  real(8) :: fblss
   !! KIT blanket model: steel fraction of breeding zone
 
-  real(dp) :: fdiv
+  real(8) :: fdiv
   !! Solid angle fraction taken by one divertor
 
-  real(dp) :: fhcd
+  real(8) :: fhcd
   !! area fraction covered by heating/current drive apparatus plus diagnostics
 
-  real(dp) :: fhole
+  real(8) :: fhole
   !! area fraction taken up by other holes (IFE)
 
   integer :: fwbsshape
@@ -60,19 +60,19 @@ module fwbs_variables
   !! - =2 defined by two ellipses
   !#TODO: change to adopt switch naming convention
 
-  real(dp) :: fwlife
+  real(8) :: fwlife
   !! first wall full-power year lifetime (y)
 
-  real(dp) :: fwmass
+  real(8) :: fwmass
   !! first wall mass (kg)
 
-  real(dp) :: fw_armour_mass
+  real(8) :: fw_armour_mass
   !! first wall armour mass (kg)
 
-  real(dp) :: fw_armour_thickness
+  real(8) :: fw_armour_thickness
   !! first wall armour thickness (m)
 
-  real(dp) :: fw_armour_vol
+  real(8) :: fw_armour_vol
   !! first wall armour volume (m3)
 
   integer :: iblanket
@@ -95,78 +95,78 @@ module fwbs_variables
   !! - =0 Frances Fox model (default)
   !! - =1 Fixed by user (qnuc)
 
-  real(dp) :: qnuc
+  real(8) :: qnuc
   !! nuclear heating in the coils (W) (`inuclear=1`)
 
-  real(dp) :: li6enrich
+  real(8) :: li6enrich
   !! lithium-6 enrichment of breeding material (%)
 
-  real(dp) :: pnucblkt
+  real(8) :: pnucblkt
   !! nuclear heating in the blanket (MW)
 
-  real(dp) :: pnuc_cp
+  real(8) :: pnuc_cp
   !! Total nuclear heating in the ST centrepost (MW)
   
-  real(dp) :: pnuc_cp_sh
+  real(8) :: pnuc_cp_sh
   !! Neutronic shield nuclear heating in the ST centrepost (MW)
   
-  real(dp) :: pnuc_cp_tf
+  real(8) :: pnuc_cp_tf
   !! TF neutronic nuclear heating in the ST centrepost (MW)
 
-  real(dp) :: pnucdiv
+  real(8) :: pnucdiv
   !! nuclear heating in the divertor (MW)
 
-  real(dp) :: pnucfw
+  real(8) :: pnucfw
   !! nuclear heating in the first wall (MW)
 
-  real(dp) :: pnuchcd
+  real(8) :: pnuchcd
   !! nuclear heating in the HCD apparatus and diagnostics (MW)
 
-  real(dp) :: pnucloss
+  real(8) :: pnucloss
   !! nuclear heating lost via holes (MW)
 
-  real(dp) :: pnucvvplus
+  real(8) :: pnucvvplus
   !! nuclear heating to vacuum vessel and beyond(MW)
 
-  real(dp) :: pnucshld
+  real(8) :: pnucshld
   !! nuclear heating in the shield (MW)
 
-  real(dp) :: whtblkt
+  real(8) :: whtblkt
   !! mass of blanket (kg)
 
-  real(dp) :: whtblss
+  real(8) :: whtblss
   !! mass of blanket - steel part (kg)
 
-  real(dp) :: armour_fw_bl_mass
+  real(8) :: armour_fw_bl_mass
   !! Total mass of armour, first wall and blanket (kg)
 
   ! CCFE HCPB Blanket Model (with or without TBR calculation) iblanket=1,3
   ! ----------
 
-  real(dp) :: breeder_f
+  real(8) :: breeder_f
   !! Volume ratio: Li4SiO4/(Be12Ti+Li4SiO4) (`iteration variable 108`)
   
-  real(dp) :: breeder_multiplier
+  real(8) :: breeder_multiplier
   !! combined breeder/multipler fraction of blanket by volume
   
-  real(dp) :: vfcblkt
+  real(8) :: vfcblkt
   !! He coolant fraction of blanket by volume (`iblanket= 1,3` (CCFE HCPB))
   
-  real(dp) :: vfpblkt
+  real(8) :: vfpblkt
   !! He purge gas fraction of blanket by volume (`iblanket= 1,3` (CCFE HCPB))
 
-  real(dp) :: whtblli4sio4
+  real(8) :: whtblli4sio4
   !! mass of lithium orthosilicate in blanket (kg) (`iblanket=1,3` (CCFE HCPB))
   
-  real(dp) :: whtbltibe12
+  real(8) :: whtbltibe12
   !! mass of titanium beryllide in blanket (kg) (`iblanket=1,3` (CCFE HCPB))
 
-  real(dp) :: neut_flux_cp
+  real(8) :: neut_flux_cp
   !! Centrepost TF fast neutron flux (E > 0.1 MeV) [m^(-2).^(-1)]
   !! This variable is only calculated for superconducting (i_tf_sup = 1 )
   !! spherical tokamal magnet designs (itart = 0)
 
-  real(dp) :: f_neut_shield
+  real(8) :: f_neut_shield
   !! Fraction of nuclear power shielded before the CP magnet (ST)
   !! ( neut_absorb = -1 --> a fit on simplified MCNP neutronic
   !! calculation is used assuming water cooled (13%) tungesten carbyde )
@@ -181,25 +181,25 @@ module fwbs_variables
   !! - =2 Lithium methatitanate
   !! - =3 Lithium zirconate
 
-  real(dp) :: densbreed
+  real(8) :: densbreed
   !! density of breeder material (kg/m3) (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: fblbe
+  real(8) :: fblbe
   !! beryllium fraction of blanket by volume (if `iblanket=2`, is Be fraction of breeding zone)
 
-  real(dp) :: fblbreed
+  real(8) :: fblbreed
   !! breeder fraction of blanket breeding zone by volume (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: fblhebmi
+  real(8) :: fblhebmi
   !! helium fraction of inboard blanket box manifold by volume (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: fblhebmo
+  real(8) :: fblhebmo
   !! helium fraction of outboard blanket box manifold by volume (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: fblhebpi
+  real(8) :: fblhebpi
   !! helium fraction of inboard blanket back plate by volume (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: fblhebpo
+  real(8) :: fblhebpo
   !! helium fraction of outboard blanket back plate by volume (`iblanket=2` (KIT HCPB))
   
   integer :: hcdportsize
@@ -209,7 +209,7 @@ module fwbs_variables
   !! - =2 'large'
   !#TODO: switch name and also large and small not descriptive enough
   
-  real(dp) :: nflutf
+  real(8) :: nflutf
   !! peak fast neutron fluence on TF coil superconductor (n/m2) (`iblanket=2` (KIT HCPB))
 
   integer :: npdiv
@@ -221,23 +221,23 @@ module fwbs_variables
   integer :: nphcdout
   !! number of outboard ports for heating/current drive (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: tbr
+  real(8) :: tbr
   !! tritium breeding ratio (`iblanket=2,3` (KIT HCPB/HCLL))
 
-  real(dp) :: tritprate
+  real(8) :: tritprate
   !! tritium production rate (g/day) (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: vvhemax
+  real(8) :: vvhemax
   !! maximum helium concentration in vacuum vessel at end of plant life (appm) 
   !! (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: wallpf
+  real(8) :: wallpf
   !! neutron wall load peaking factor (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: whtblbreed 
+  real(8) :: whtblbreed 
   !! mass of blanket - breeder part (kg) (`iblanket=2` (KIT HCPB))
 
-  real(dp) :: whtblbe
+  real(8) :: whtblbe
   !! mass of blanket - beryllium part (kg)
 
   ! CCFE HCPB model with Tritium Breeding Ratio calculation (iblanket=3)
@@ -275,11 +275,11 @@ module fwbs_variables
   !! - =2 pressurized water
   !#TODO: change switch name to satisfy convention
 
-  real(dp) :: afwi
+  real(8) :: afwi
   !! inner radius of inboard first wall/blanket coolant channels (stellarator only) (m)
   !#TODO move to stellarator?
 
-  real(dp) :: afwo
+  real(8) :: afwo
   !! inner radius of outboard first wall/blanket coolant channels (stellarator only) (m)
   !#TODO move to stellarator?
 
@@ -289,52 +289,52 @@ module fwbs_variables
   !! - 'helium' 
   !! - 'water'
 
-  real(dp) :: fw_wall
+  real(8) :: fw_wall
   !! wall thickness of first wall coolant channels (m)
 
-  real(dp) :: afw
+  real(8) :: afw
   !! radius of first wall cooling channels (m)
   
-  real(dp) :: pitch
+  real(8) :: pitch
   !! pitch of first wall cooling channels (m)
 
-  real(dp) :: fwinlet
+  real(8) :: fwinlet
   !! inlet temperature of first wall coolant (K)
 
-  real(dp) :: fwoutlet
+  real(8) :: fwoutlet
   !! outlet temperature of first wall coolant (K)
 
-  real(dp) :: fwpressure
+  real(8) :: fwpressure
   !! first wall coolant pressure (Pa) (`secondary_cycle>1`)
 
-  real(dp) :: tpeak
+  real(8) :: tpeak
   !! peak first wall temperature (K)
 
-  real(dp) :: roughness
+  real(8) :: roughness
   !! first wall channel roughness epsilon (m)
 
-  real(dp) :: fw_channel_length
+  real(8) :: fw_channel_length
   !! Length of a single first wall channel (all in parallel) (m)
   !! (`iteration variable 114`, useful for `constraint equation 39`)
 
-  real(dp) :: peaking_factor
+  real(8) :: peaking_factor
   !! peaking factor for first wall heat loads. (Applied separately to inboard and outboard loads.
   !! Applies to both neutron and surface loads. Only used to calculate peak temperature - not 
   !! the coolant flow rate.)
 
-  real(dp) :: blpressure
+  real(8) :: blpressure
   !! blanket coolant pressure (Pa) (`secondary_cycle>1`)
 
-  real(dp) :: inlet_temp
+  real(8) :: inlet_temp
   !! inlet temperature of blanket coolant  (K) (`secondary_cycle>1`)
 
-  real(dp) :: outlet_temp
+  real(8) :: outlet_temp
   !! Outlet temperature of blanket coolant (K) (`secondary_cycle>1`)
   !!
   !! - input if `coolwh=1` (helium)
   !! - calculated if `coolwh=2` (water)
 
-  real(dp) :: coolp
+  real(8) :: coolp
   !! blanket coolant pressure (Pa) (stellarator only)
 
   integer :: nblktmodpo
@@ -349,117 +349,117 @@ module fwbs_variables
   integer :: nblktmodti
   !! number of inboard blanket modules in toroidal direction (`secondary_cycle>1`)
 
-  real(dp) :: tfwmatmax
+  real(8) :: tfwmatmax
   !! maximum temperature of first wall material (K) (`secondary_cycle>1`)
 
-  real(dp) :: fw_th_conductivity
+  real(8) :: fw_th_conductivity
   !! thermal conductivity of first wall material at 293 K (W/m/K) (Temperature dependence 
   !! is as for unirradiated Eurofer)
 
-  real(dp) :: fvoldw
+  real(8) :: fvoldw
   !! area coverage factor for vacuum vessel volume
 
-  real(dp) :: fvolsi
+  real(8) :: fvolsi
   !! area coverage factor for inboard shield volume
 
-  real(dp) :: fvolso
+  real(8) :: fvolso
   !! area coverage factor for outboard shield volume
 
-  real(dp) :: fwclfr
+  real(8) :: fwclfr
   !! first wall coolant fraction (calculated if `lpulse=1` or `ipowerflow=1`)
 
-  real(dp) :: praddiv
+  real(8) :: praddiv
   !! Radiation power incident on the divertor (MW)
 
-  real(dp) :: pradfw
+  real(8) :: pradfw
   !! Radiation power incident on the first wall (MW)
 
-  real(dp) :: pradhcd
+  real(8) :: pradhcd
   !! Radiation power incident on the heating and current drive system (MW)
 
-  real(dp) :: pradloss
+  real(8) :: pradloss
   !! Radiation power lost through holes (eventually hits shield) (MW)
   !! Only used for stellarator
 
-  real(dp) :: ptfnuc
+  real(8) :: ptfnuc
   !! nuclear heating in the TF coil (MW)
 
-  real(dp) :: ptfnucpm3
+  real(8) :: ptfnucpm3
   !! nuclear heating in the TF coil (MW/m3) (`blktmodel>0`)
   !#TODO: check usage of old blktmodel. Update to iblanket
 
-  real(dp) :: rdewex
+  real(8) :: rdewex
   !! cryostat radius (m)
 
-  real(dp) :: zdewex
+  real(8) :: zdewex
   !! cryostat height (m)
 
-  real(dp) :: rpf2dewar
+  real(8) :: rpf2dewar
   !! radial distance between outer edge of largest (`ipfloc=3`) PF coil (or stellarator 
   !! modular coil) and cryostat (m)
 
-  real(dp) :: vdewex
+  real(8) :: vdewex
   !! cryostat volume (m3)
 
-  real(dp) :: vdewin
+  real(8) :: vdewin
   !! vacuum vessel volume (m3)
 
-  real(dp) :: vfshld
+  real(8) :: vfshld
   !! coolant void fraction in shield
 
-  real(dp) :: volblkt
+  real(8) :: volblkt
   !! volume of blanket (m3)
 
-  real(dp) :: volblkti
+  real(8) :: volblkti
   !! volume of inboard blanket (m3)
 
-  real(dp) :: volblkto
+  real(8) :: volblkto
   !! volume of outboard blanket (m3)
 
-  real(dp) :: volshld
+  real(8) :: volshld
   !! volume of shield (m3)
 
-  real(dp) :: whtshld
+  real(8) :: whtshld
   !! mass of shield (kg)
 
-  real(dp) :: wpenshld
+  real(8) :: wpenshld
   !! mass of the penetration shield (kg)
 
-  real(dp) :: wtshldi
+  real(8) :: wtshldi
   !! mass of inboard shield (kg)
 
-  real(dp) :: wtshldo
+  real(8) :: wtshldo
   !! mass of outboard shield (kg)
   
   integer :: irefprop
   !! Switch to use REFPROP routines (stellarator only)
   !#TODO: number of stellarator only items here. Also appear in fispact. Tidy needed
 
-  real(dp) :: fblli
+  real(8) :: fblli
   !! lithium fraction of blanket by volume (stellarator only)
 
-  real(dp) :: fblli2o
+  real(8) :: fblli2o
   !! lithium oxide fraction of blanket by volume (stellarator only)
 
-  real(dp) :: fbllipb
+  real(8) :: fbllipb
   !! lithium lead fraction of blanket by volume (stellarator only)
 
-  real(dp) :: fblvd
+  real(8) :: fblvd
   !! vanadium fraction of blanket by volume (stellarator only)
 
-  real(dp) :: wtblli2o
+  real(8) :: wtblli2o
   !! mass of blanket - Li_2O part (kg)
 
-  real(dp) :: wtbllipb
+  real(8) :: wtbllipb
   !! mass of blanket - Li-Pb part (kg)
   
-  real(dp) :: whtblvd
+  real(8) :: whtblvd
   !! mass of blanket - vanadium part (kg)
 
-  real(dp) :: whtblli
+  real(8) :: whtblli
   !! mass of blanket - lithium part (kg)
 
-  real(dp) :: vfblkt
+  real(8) :: vfblkt
   !! coolant void fraction in blanket (`blktmodel=0`), (calculated if `blktmodel > 0`)
 
   integer :: blktmodel
@@ -469,13 +469,13 @@ module fwbs_variables
   !! - =1 KIT model based on a helium-cooled pebble-bed blanket (HCPB) reference design
   !#TODO: this needs investigating and removing after any required functionality is in iblanket
 
-  real(dp) :: declblkt
+  real(8) :: declblkt
   !! neutron power deposition decay length of blanket structural material (m) (stellarators only)
 
-  real(dp) :: declfw
+  real(8) :: declfw
   !! neutron power deposition decay length of first wall structural material (m) (stellarators only)
 
-  real(dp) :: declshld
+  real(8) :: declshld
   !! neutron power deposition decay length of shield structural material (m) (stellarators only)
 
   integer :: blkttype
@@ -486,10 +486,10 @@ module fwbs_variables
   !! - =3 HCPB; efficiency taken from WP12-DAS08-T01, EFDA_D_2LLNBX
   !#TODO: this needs to be merged into iblanket and then removed.
 
-  real(dp) :: etaiso
+  real(8) :: etaiso
   !! isentropic efficiency of FW and blanket coolant pumps
 
-  real(dp) :: etahtp
+  real(8) :: etahtp
   !! electrical efficiency of primary coolant pumps
 
   contains
